refactor: Extract converter and property logic to separate handlers
This commit extracts the converter and property-related logic from `JsonMapper` into new `ConverterHandler` and `PropertyHandler` classes. This improves the separation of concerns and makes the `JsonMapper` class smaller and more focused.

The following changes were made:
- Created a new `ConverterHandler` class in `mapper/lib/src/logic/converter_handler.dart`.
- Moved `getConverter`, `getEnumDescriptor`, `getConvertedValue`, and `configureConverter` methods from `JsonMapper` to `ConverterHandler`.
- Moved `converters` and `convertedValuesCache` state from `JsonMapper` to `ConverterHandler`.
- Created a new `PropertyHandler` class in `mapper/lib/src/logic/property_handler.dart`.
- Moved `enumeratePublicProperties` and `resolveProperty` methods from `JsonMapper` to `PropertyHandler`.
- Updated all call sites to use the new handlers.
- Removed unused imports from modified files.
